scraper/requirements.txt
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
jinja2==3.1.4
python-dotenv==1.0.1
